Refactor database models: replace Article, User, and Session schemas with new implementations in 321.js and 666.js; update import paths in service and middleware files.
