+++ src/services/auth/logoutUserService.js
@@ -1,4 +1,4 @@
-import { Session } from '../../db/models/Session.js';
+import { Session } from '../../db/models/123.js';
 
 export const logoutUser = async (sessionId, sessionToken) => {
   const deletedSession = await Session.findOneAndDelete({
